+++ src/main/java/urjc/isi/dao/interfaces/PeliculasDAO.java
@@ -22,5 +22,5 @@ public interface PeliculasDAO extends GenericDAO<Peliculas>{
 
 	public List<Peliculas> selectAllWhereGuionista(String name);
 	
-	public List<Peliculas> selectAllWhereDuration(int value);
+	public List<Peliculas> selectAllWhereDuration(float value);
 }
